+++ src/frontend/src/features/rooms/api/ApiRoom.ts
@@ -33,6 +33,7 @@ export type ApiRoom = {
   pin_code?: string
   is_administrable: boolean
   access_level: ApiAccessLevel
+  access_level_overridden?: boolean
   livekit?: ApiLiveKit
   configuration?: RoomConfiguration
   /**